refactor: split statement parser utilities
