+++ src/parser/stmt/mod.rs
@@ -9,26 +9,32 @@
 //! - Recovery stops at PHP statement boundaries so follow-up diagnostics remain useful.
 
 mod assign;
+mod blocks;
 mod declare;
 mod ffi;
+mod names;
 mod namespace_use;
 mod oop;
 /// Parameter parsing helpers for function declarations, including typed parameters and return types.
 pub(crate) mod params;
 mod simple;
+mod recovery;
 
 use crate::errors::CompileError;
 use crate::lexer::{SpannedToken, Token};
-use crate::names::{Name, NameKind};
 use crate::parser::ast::{AttributeGroup, ExprKind, Stmt, StmtKind};
 use crate::parser::control;
 use crate::parser::expr::parse_expr;
 use crate::span::Span;
 
 pub use ffi::parse_extern_stmts;
+pub use blocks::{parse_block, parse_body};
 pub(crate) use oop::parse_anonymous_class;
 pub(crate) use params::{looks_like_typed_param, parse_type_expr};
 pub(crate) use assign::can_replay_assignment_target;
+pub(crate) use blocks::{expect_semicolon, expect_token};
+pub(crate) use names::{name_part_from_token, name_starts_at, parse_name, parse_unqualified_name};
+pub(crate) use recovery::recover_to_statement_boundary;
 
 /// Parses a single PHP statement, including optional PHP 8 attribute groups.
 pub fn parse_stmt(tokens: &[SpannedToken], pos: &mut usize) -> Result<Stmt, CompileError> {
@@ -289,288 +295,4 @@ fn statement_lhs_contains_double_colon(tokens: &[SpannedToken], start: usize) ->
         }
     }
     false
-}
-
-/// Advances `pos` to the next PHP statement boundary following a parse error.
-///
-/// Skips tokens while tracking parenthesis and bracket depth, stopping at `;`, `}`, or `EOF`
-/// when depth is zero. Also stops when a statement-starting keyword is encountered (after the
-/// first token), allowing recovery to continue from the next statement without consuming it.
-/// If already at a boundary token, advances by one to avoid infinite loops.
-pub(crate) fn recover_to_statement_boundary(tokens: &[SpannedToken], pos: &mut usize) {
-    let start = *pos;
-    let mut paren_depth = 0usize;
-    let mut bracket_depth = 0usize;
-
-    while *pos < tokens.len() {
-        match tokens[*pos].0 {
-            Token::LParen => {
-                paren_depth += 1;
-                *pos += 1;
-            }
-            Token::RParen => {
-                paren_depth = paren_depth.saturating_sub(1);
-                *pos += 1;
-            }
-            Token::LBracket => {
-                bracket_depth += 1;
-                *pos += 1;
-            }
-            Token::RBracket => {
-                bracket_depth = bracket_depth.saturating_sub(1);
-                *pos += 1;
-            }
-            Token::Semicolon if paren_depth == 0 && bracket_depth == 0 => {
-                *pos += 1;
-                break;
-            }
-            Token::RBrace if paren_depth == 0 && bracket_depth == 0 => {
-                break;
-            }
-            Token::EndDeclare if paren_depth == 0 && bracket_depth == 0 => {
-                break;
-            }
-            Token::Eof if paren_depth == 0 && bracket_depth == 0 => {
-                break;
-            }
-            Token::Echo
-            | Token::Print
-            | Token::Variable(_)
-            | Token::This
-            | Token::PlusPlus
-            | Token::MinusMinus
-            | Token::Class
-            | Token::Enum
-            | Token::ReadOnly
-            | Token::Packed
-            | Token::Interface
-            | Token::Trait
-            | Token::Abstract
-            | Token::Final
-            | Token::Function
-            | Token::Namespace
-            | Token::Use
-            | Token::Declare
-            | Token::Return
-            | Token::Throw
-            | Token::Include
-            | Token::IncludeOnce
-            | Token::Require
-            | Token::RequireOnce
-            | Token::Const
-            | Token::Global
-            | Token::Static
-            | Token::Identifier(_)
-            | Token::Self_
-            | Token::Parent
-            | Token::Backslash
-            | Token::Question
-            | Token::Switch
-            | Token::If
-            | Token::IfDef
-            | Token::Try
-            | Token::While
-            | Token::Do
-            | Token::For
-            | Token::Foreach
-            | Token::Break
-            | Token::Continue
-                if *pos > start && paren_depth == 0 && bracket_depth == 0 =>
-            {
-                break;
-            }
-            _ => {
-                *pos += 1;
-            }
-        }
-    }
-
-    if *pos == start && *pos < tokens.len() && !matches!(tokens[*pos].0, Token::Eof) {
-        *pos += 1;
-    }
-}
-
-/// Parses a braced block `{ stmts }`, returning statements or errors.
-pub fn parse_block(tokens: &[SpannedToken], pos: &mut usize) -> Result<Vec<Stmt>, CompileError> {
-    let span = if *pos < tokens.len() {
-        tokens[*pos].1.span
-    } else {
-        Span::dummy()
-    };
-    expect_token(tokens, pos, &Token::LBrace, "Expected '{'")?;
-
-    let mut stmts = Vec::new();
-    let mut errors = Vec::new();
-    while *pos < tokens.len() && !matches!(tokens[*pos].0, Token::RBrace | Token::Eof) {
-        match parse_stmt(tokens, pos) {
-            Ok(stmt) => stmts.push(stmt),
-            Err(error) => {
-                errors.extend(error.flatten());
-                recover_to_statement_boundary(tokens, pos);
-            }
-        }
-    }
-
-    if *pos >= tokens.len() || tokens[*pos].0 != Token::RBrace {
-        errors.push(CompileError::new(span, "Expected '}'"));
-        return Err(CompileError::from_many(errors));
-    }
-    *pos += 1;
-
-    if errors.is_empty() {
-        Ok(stmts)
-    } else {
-        Err(CompileError::from_many(errors))
-    }
-}
-
-/// Parse either a braced block `{ ... }` or a single statement (for braceless if/while/for/foreach).
-pub fn parse_body(tokens: &[SpannedToken], pos: &mut usize) -> Result<Vec<Stmt>, CompileError> {
-    if *pos < tokens.len() && tokens[*pos].0 == Token::LBrace {
-        parse_block(tokens, pos)
-    } else {
-        let stmt = parse_stmt(tokens, pos)?;
-        Ok(vec![stmt])
-    }
-}
-
-/// Consumes a semicolon token, or returns an error if the current token is not a semicolon.
-///
-/// Used to terminate expression statements, return statements, throw statements, and similar
-/// constructs that require explicit semicolons in PHP.
-pub(crate) fn expect_semicolon(
-    tokens: &[SpannedToken],
-    pos: &mut usize,
-) -> Result<(), CompileError> {
-    if *pos < tokens.len() && tokens[*pos].0 == Token::Semicolon {
-        *pos += 1;
-        Ok(())
-    } else {
-        let span = if *pos < tokens.len() {
-            tokens[*pos].1.span
-        } else {
-            Span::dummy()
-        };
-        Err(CompileError::new(span, "Expected ';'"))
-    }
-}
-
-/// Consumes the expected token, advancing `pos`, or returns an error with the given message.
-///
-/// Used for mandatory syntax elements like `{` in blocks or specific keywords where absence
-/// indicates a syntax error.
-pub(crate) fn expect_token(
-    tokens: &[SpannedToken],
-    pos: &mut usize,
-    expected: &Token,
-    msg: &str,
-) -> Result<(), CompileError> {
-    if *pos < tokens.len() && tokens[*pos].0 == *expected {
-        *pos += 1;
-        Ok(())
-    } else {
-        let span = if *pos < tokens.len() {
-            tokens[*pos].1.span
-        } else {
-            Span::dummy()
-        };
-        Err(CompileError::new(span, msg))
-    }
-}
-
-/// Converts a token accepted as an ordinary PHP name segment to its source spelling.
-///
-/// `enum` is a soft keyword in class-like name contexts; other reserved words remain
-/// excluded here even though PHP permits them in the broader member-name grammar.
-pub(crate) fn name_part_from_token(
-    token: &Token,
-    metadata: &crate::lexer::TokenMetadata,
-) -> Option<String> {
-    match token {
-        Token::Identifier(name) => Some(name.clone()),
-        Token::Enum => crate::parser::keyword_name::bareword_name_from_token(token, metadata),
-        _ => None,
-    }
-}
-
-/// Returns true if the token at `pos` starts a PHP class-like name.
-///
-/// Accepts identifiers, the soft keyword `enum`, and a leading namespace separator.
-pub(crate) fn name_starts_at(tokens: &[SpannedToken], pos: usize) -> bool {
-    match tokens.get(pos) {
-        Some((Token::Backslash, _)) => true,
-        Some((token, metadata)) => name_part_from_token(token, metadata).is_some(),
-        None => false,
-    }
-}
-
-/// Parses one unqualified class-like declaration name.
-///
-/// This accepts the soft keyword `enum` alongside ordinary identifiers but never
-/// consumes namespace separators, which are invalid in declaration names.
-pub(crate) fn parse_unqualified_name(
-    tokens: &[SpannedToken],
-    pos: &mut usize,
-    span: Span,
-    error: &str,
-) -> Result<String, CompileError> {
-    let Some((token, metadata)) = tokens.get(*pos) else {
-        return Err(CompileError::new(span, error));
-    };
-    let name = name_part_from_token(token, metadata)
-        .ok_or_else(|| CompileError::new(span, error))?;
-    *pos += 1;
-    Ok(name)
-}
-
-/// Parses a PHP qualified or unqualified name from the token stream.
-///
-/// Handles backslash-prefixed fully-qualified names (`\Foo\Bar`), qualified names (`Foo\Bar`),
-/// and simple identifiers (`Foo`). Sets `NameKind` based on leading backslash presence and
-/// tracks whether any intermediate backslashes appeared. Returns an error if no identifier
-/// is found or if a backslash appears at the end of the sequence.
-pub(crate) fn parse_name(
-    tokens: &[SpannedToken],
-    pos: &mut usize,
-    span: Span,
-    first_error: &str,
-) -> Result<Name, CompileError> {
-    let mut kind = NameKind::Unqualified;
-    if *pos < tokens.len() && tokens[*pos].0 == Token::Backslash {
-        kind = NameKind::FullyQualified;
-        *pos += 1;
-    }
-
-    let mut parts = Vec::new();
-    loop {
-        match tokens.get(*pos) {
-            Some((token, metadata)) if name_part_from_token(token, metadata).is_some() => {
-                parts.push(
-                    name_part_from_token(token, metadata)
-                        .expect("name part was checked immediately above"),
-                );
-                *pos += 1;
-            }
-            _ if parts.is_empty() => {
-                return Err(CompileError::new(span, first_error))
-            }
-            _ => {
-                return Err(CompileError::new(
-                    span,
-                    "Expected identifier after '\\' in qualified name",
-                ))
-            }
-        }
-
-        if *pos < tokens.len() && tokens[*pos].0 == Token::Backslash {
-            if kind != NameKind::FullyQualified {
-                kind = NameKind::Qualified;
-            }
-            *pos += 1;
-            continue;
-        }
-        break;
-    }
-
-    Ok(Name::from_parts(kind, parts))
 }
